Update HCloud.py
Fixed - During the dependency install for pywin32 on Windows the code performs the post-install and fails to verify. But the script just needed to be closed and ran again.  So the new version of the code will restart itself to finalize the pywin32 installation and successfully open the GUI window.

Fixed - If the .ssh folder did not already exist then the code was failing and causing it to go no further when it was checking for local ssh files to populate in the dropdown menu.  Now the script will recognize if the folder does not exist and know not to try to list any local ssh files in the dropdown.