Update HCloud.py
Added: When a new cloud server is created a ssh_config file is created in a SERVERS/$servername/$servername_ssh_config.txt file. This file holds all of the configuration for the server. It also has the ssh and sftp command lines for the user to access the server. The ssh_config file can also be used to import the cloud server configuration into Termius.

Added: A checkbox on the Install nodectl tab to enable/disable shortcuts creation on the desktop, for easy access to the cloud server.